# Opertaion of selection
In the connect_to_database function, you should use return conn, cur instead of return conn, cursor.

In the display_sorted_cities function, you are using orders as a parameter, but you are checking for order inside the function. Also, you are using order instead of orders in the if conditions. Fix that.

In the display_highest_population and display_lowest_population_city functions, you are using undefined variables like highest_population_city and lowest_population_city. Use highest_population and lowest_population instead.

In the main function, the indentation for the user input section is incorrect. It should be inside the while True loop.

Use conn.commit() to commit changes after executing queries that modify the database.